// miniprogram/miniprogram/types/profile-overview.ts
export interface AppUserOverview {
  availableCouponCount: number;
  favoriteCount: number;
  browseHistoryCount: number;
  unpaidOrderCount: number;
  toShipOrderCount: number;
  toReceiveOrderCount: number;
  toReviewOrderCount: number;
  activeAfterSaleCount: number;
  customerServiceUnreadCount: number;
  customerServiceOnline: boolean;
}